Update utils.py
Changed the format of the date when saving the snaps. Updated it to fit the ISO 8601 standard ( https://en.wikipedia.org/wiki/ISO_8601 ).
Also, added the format of saving to date - user, as it is easier to read in a folder. ( It is also easier to grep all of the pics of a user than to get the time ).